+++ taf/__init__.py
@@ -1,5 +1,6 @@
 """Add platform-dependent libraries to the path.
 """
+
 import sys