+++ taf/updater/updater_pipeline.py
@@ -2,6 +2,7 @@ from collections import defaultdict
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from enum import Enum
 import functools
+import json
 from pathlib import Path
 import re
 import shutil
@@ -314,6 +315,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     RunMode.UPDATE,
                     self.should_update_auth_repos,
                 ),  # auth repo
+                (self.set_excluded_targets, RunMode.ALL, self.should_run_step_default),
                 # should_validate_target_repos
                 (
                     self.load_target_repositories,
@@ -426,7 +428,9 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
         )
         self.checkout = update_config.checkout
         self.bare = update_config.bare
-        self.excluded_target_globs = update_config.excluded_target_globs
+        self.excluded_target_globs = update_config.excluded_target_globs or []
+        self.exclude_filter = update_config.exclude_filter
+        self.sync_all = update_config.sync_all
         self.no_targets = update_config.no_targets
         self.no_upstream = update_config.no_upstream
         self.force = update_config.force
@@ -1103,6 +1107,44 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                 last_commits_per_repos=last_commits_per_repos,
             )
         )
+
+    def set_excluded_targets(self):
+        if self.sync_all:
+            self.excluded_target_globs = []
+        else:
+            self.excluded_target_globs = list(self.excluded_target_globs)
+            if self.exclude_filter:
+                excluded_repo_names = repositoriesdb.get_repository_names_by_expression(
+                    self.state.users_auth_repo, filter_expr=self.exclude_filter
+                )
+                if excluded_repo_names:
+                    self.excluded_target_globs.extend(excluded_repo_names)
+
+            last_validated_data = self.state.users_auth_repo.last_validated_data
+            if last_validated_data:
+                if len(last_validated_data) == 1:
+                    # check if old last validated format
+                    # not a json file, just one commit
+                    last_validated_file_content = (
+                        self.state.users_auth_repo.get_last_validated_file_content()
+                    )
+                    try:
+                        json.loads(last_validated_file_content)
+                    except json.decoder.JSONDecodeError:
+                        return UpdateStatus.SUCCESS
+
+                all_repositories = repositoriesdb.load_repositories_json(
+                    self.state.users_auth_repo
+                )
+                if all_repositories is not None:
+                    all_repositories = all_repositories.get("repositories")
+                    skipped_repositories = [
+                        repository
+                        for repository in all_repositories
+                        if repository not in last_validated_data
+                    ]
+                    self.excluded_target_globs.extend(skipped_repositories)
+            return UpdateStatus.SUCCESS
 
     def load_target_repositories(self):
         taf_logger.debug(f"{self.state.auth_repo_name}: Loading target repositories...")