+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Added
 
+- Add Python expression filtering for repository exclusion ([694])
 - Git: add branch fast forward ([#691])
 
 ### Changed
@@ -19,7 +20,8 @@ and this project adheres to [Semantic Versioning][semver].
 - Fix `commit_exists` ([689])
 - Attempt to remove `index.lock` if branch checkout fails ([684])
 
-[693]: https://github.com/openlawlibrary/taf/pull/6930
+[694]: https://github.com/openlawlibrary/taf/pull/694
+[693]: https://github.com/openlawlibrary/taf/pull/693
 [691]: https://github.com/openlawlibrary/taf/pull/691
 [689]: https://github.com/openlawlibrary/taf/pull/689
 [684]: https://github.com/openlawlibrary/taf/pull/684